least_squares moved into newly created objective folder and core renamed to _core and top level imports created for important classes and functions
