+++ examples/eliminate_magnetic_islands
@@ -13,8 +13,8 @@ to the approach of Hanson and Cary (1984).
 
 import numpy as np
 from simsopt.mhd.spec import Spec, Residue
-from simsopt.core.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
+from simsopt.objectives.least_squares import LeastSquaresProblem
+from simsopt.solve.serial import least_squares_serial_solve
 
 # Initialze a Spec object from a standard SPEC input file:
 s = Spec('QH-residues.sp')
