+++ examples/minimize_curve_length
@@ -3,9 +3,9 @@
 import numpy as np
 from simsopt.geo.curverzfourier import CurveRZFourier
 from simsopt.geo.curveobjectives import CurveLength
-from simsopt.core.optimizable import optimizable
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
+from simsopt import optimizable
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_serial_solve
 
 """
 Minimize the length of a curve, holding the 0-frequency Fourier mode fixed.
