+++ examples/qsc
@@ -3,9 +3,9 @@
 import logging
 import numpy as np
 from qsc import Qsc
-from simsopt.core.optimizable import optimizable
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
+from simsopt import optimizable
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_serial_solve
 
 """
 Optimize the Stellarator Quasisymmetry Construction code
