+++ examples/resolution_increase
@@ -19,8 +19,8 @@ import numpy as np
 
 from simsopt.util.mpi import MpiPartition, log
 from simsopt.mhd import Vmec, Boozer, Quasisymmetry
-from simsopt.core.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.mpi_solve import least_squares_mpi_solve
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_mpi_solve
 
 #log()
