+++ examples/stellopt_scenarios_1DOF_circularCrossSection_varyAxis_targetIota
@@ -4,9 +4,9 @@ from mpi4py import MPI
 import numpy as np
 
 from simsopt.mhd import Vmec
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
+from simsopt import LeastSquaresProblem
 from simsopt.util.mpi import MpiPartition, log
-from simsopt.solve.mpi_solve import least_squares_mpi_solve
+from simsopt import least_squares_mpi_solve
 
 """
 This script implements the "1DOF_circularCrossSection_varyAxis_targetIota"
