+++ examples/stellopt_scenarios_1DOF_circularCrossSection_varyR0_targetVolume
@@ -3,10 +3,10 @@
 import numpy as np
 from mpi4py import MPI
 
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
+from simsopt import LeastSquaresProblem
 from simsopt.mhd import Vmec
 from simsopt.util.mpi import MpiPartition, log
-from simsopt.solve.mpi_solve import least_squares_mpi_solve
+from simsopt import least_squares_mpi_solve
 
 """
 This script implements the "1DOF_circularCrossSection_varyR0_targetVolume"
