+++ examples/stellopt_scenarios_1DOF_circularCrossSection_varyR0_targetVolume_spec
@@ -5,8 +5,8 @@ import numpy as np
 
 from simsopt.mhd import Spec
 from simsopt.geo.surfacerzfourier import SurfaceRZFourier
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_serial_solve
 
 """
 This script implements the "1DOF_circularCrossSection_varyR0_targetVolume"
