+++ examples/stellopt_scenarios_2DOF_circularCrossSection_varyAxis_targetIotaAndQuasisymmetry
@@ -8,8 +8,8 @@ of the objective function landscape.
 """
 
 from simsopt.mhd import Vmec, Boozer, Quasisymmetry
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_serial_solve
 
 vmec = Vmec("input.2DOF_circularCrossSection_varyAxis_targetIotaAndQuasisymmetry")
