+++ examples/stellopt_scenarios_2DOF_vmecAndSpec
@@ -4,8 +4,8 @@ import logging
 import numpy as np
 
 from simsopt.mhd import Vmec, Spec
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_serial_solve
 
 """
 This script implements the "2DOF_vmecOnly_targetIotaAndVolume" example from
