+++ src/simsopt/_core/__init__.py
@@ -1,0 +1,1 @@
+from .optimizable import Optimizable, optimizable
