--- src/simsopt/core/__init__.py
@@ -1,12 +1,0 @@
-#from .util import *
-#from .optimizable import *
-#from .surface import *
-#from .functions import *
-#from .dofs import *
-#from .least_squares_problem import LeastSquaresTerm, LeastSquaresProblem
-
-# This next bit is to suppress a Jax warning:
-import warnings
-warnings.filterwarnings("ignore", message="No GPU/TPU found")
-
-#all = ['Parameter']
