+++ src/simsopt/objectives/__init__.py
@@ -1,0 +1,3 @@
+from .least_squares import LeastSquaresTerm, LeastSquaresProblem
+
+__all__ = ['LeastSquaresTerm', 'LeastSquaresProblem']
