+++ src/simsopt/solve/__init__.py
@@ -1,2 +1,5 @@
-from .serial_solve import least_squares_serial_solve, serial_solve
-from .mpi_solve import least_squares_mpi_solve, fd_jac_mpi
+from .serial import least_squares_serial_solve, serial_solve
+from .mpi import least_squares_mpi_solve, fd_jac_mpi
+
+__all__ = ['least_squares_serial_solve', 'serial_solve',
+           'least_squares_mpi_solve', 'fd_jac_mpi']
