+++ src/simsopt/util/__init__.py
@@ -1,1 +1,2 @@
 from .mpi import *
+from .logging import *
