+++ tests/core/test_integrated.py
@@ -5,11 +5,11 @@ import numpy as np
 
 from simsopt.geo.surfacerzfourier import SurfaceRZFourier
 from simsopt.geo.surfacegarabedian import SurfaceGarabedian
-from simsopt.core.optimizable import optimizable
 from simsopt.util.mpi import MpiPartition
-from simsopt.solve.least_squares_problem import LeastSquaresProblem
-from simsopt.solve.serial_solve import least_squares_serial_solve
-from simsopt.solve.mpi_solve import least_squares_mpi_solve
+from simsopt import optimizable
+from simsopt import LeastSquaresProblem
+from simsopt import least_squares_serial_solve
+from simsopt import least_squares_mpi_solve
 
 def mpi_solve_1group(prob, **kwargs):
     least_squares_mpi_solve(prob, MpiPartition(ngroups=1), **kwargs)
